Making a Main Menu and Settings Scene.
Currently working on changing the different background music and sound effects in the main menu and the level. Planning to make my custom slider and add my own UI buttons that I used in one of my other projects for the menus. Will add the background music and sound effects later on.